Oops forgot to re-uncomment code that handles deleting generated images after returning them (waas testing chart gen changes with local images).
For future reference, this results in just two files in the temp_renders folder per-device and per-chart (so 4 files total for swell and tide and jpeg and raw). It doesn't build up forever because the filename is generated from chart type and device ID, no state included. Worth remembering for the future for diagnosing any issues